// feature-libs/product/configuration/src/textfield-configurator/model/configurator-textfield.model.ts
import { GenericConfigurator } from '@spartacus/core';

export namespace ConfiguratorTextfield {
  export interface Configuration {
    configurationInfos: ConfigurationInfo[];
    owner?: GenericConfigurator.Owner;
  }

  export interface ConfigurationInfo {
    configurationLabel?: string;
    configurationValue?: string;
    status?: string;
  }

  export interface AddToCartParameters {
    userId: string;
    cartId: string;
    productCode: string;
    quantity: number;
    configuration?: Configuration;
  }

  export interface UpdateCartEntryParameters {
    userId: string;
    cartId: string;
    cartEntryNumber: string;
    configuration?: Configuration;
  }
}